Extract generally useful property event handlers
Those are likely to be handy for tests.
